netrc/netrc.go: use constant format string
The "go vet -printf ./..." command, when run with Go 1.24 (and a matching
Go version number in the "go.mod" file), now detects format strings that
are not constants, as described in golang/go#60529.

We have one such non-constant format string in the newToken() function
of the "netrc" package, so we adjust it now to use a standard format
string with a specifier verb.  This issue was reported in:

  https://github.com/git-lfs/git-lfs/issues/5968#issuecomment-2600792917